Move methods to the right util file
