+++ src/nl/rubensten/texifyidea/intentions/LatexMoveSelectionToFileIntention.kt
@@ -64,13 +64,12 @@ open class LatexMoveSelectionToFileIntention : TexifyIntentionBase("Move selecti
         @Language("RegExp")
         val fileName = dialogResult.replace(Regex("(\\.tex)+$", RegexOption.IGNORE_CASE), "")
         // Note that we do not override the user-specified filename to be LaTeX-like.
-
         val root = file.findRootFile().containingDirectory.virtualFile.canonicalPath
 
         // Execute write actions.
         runWriteAction {
             val filePath = "$root/$fileName.tex"
-            val createdFile = TexifyUtil.createFile(filePath, text.toString())
+            val createdFile = createFile(filePath, text.toString())
 
             for ((start, end) in offsets.reversed()) {
                 document.deleteString(start, end)
